[#1,content/faq][s]: refactor faqs into one big markdown file as that is simplest way to get this working.
Currently had faqs in separate files so i could do fancy faq page with snippets at the top of page and full content below. This was nice but not really essential and issues like putting ordering in weight field. If we want to reintroduce this we probably want simple markdown faq files and then a page e.g. pages/faq.js that lists faqs in order and loads, excerpts then presents.